Modified attributes for the manner Anthony and I discussed in the meeting. I will be sending an email to the group on this discussion but will include it here for our records.
Attribute is now a superclass to cover different types of data that might be searched for. This data includes fields, or checkboxes. By designing it in this manner, different input can be handled differently.

FieldAttribute is the first kind to be tested and is for standard input boxes. It has a validate method which checks if the provided input is input correctly and if it is a valid searchable input.

Anthony, can you check to make sure that I utilized the constraints properly? Is this the manner we discussed? Also, how can I determine the column name based on the constraint? Is there a way to find out the constraint name, or do I need to somehow move that server side?