I apologize in advance if I broke anything. I was trying to look at how all the other service things work on the back end and mimicked those. So I created a serviceImpl and added the servlet to web.xml and MetPetDBApplication.gwt.xml.
I also changed the Search Service stuff to return a list rather than Results. Everything else returns Results but I don't know how that works or how it's created.

The Search front end is back to being empty. I'm trying to copy the others but whew, it's not getting very far. I've been trying to bite off of bulkupload's implementation but it's taking time to get into this big ol head of mine.