Added more constraints, inputs into search.
However, issues that don't work:

Owner throws an error as a constraint. It seems to have trouble with the maxlength/minlength requirements
Location doesn't work, it never finds an item with the location no matter what i put. It also makes itself required which is obnoxious, AND it names both text boxes as location.
Using collection date seems to require precision. I'm not sure how to make this user inputted, will require further formatting of the input box.
Rock type needs to be changed to a different style of input, one that allows a set of data to be chosen. Do we have any precedents of this?