Created a CheckBoxesAttribute
However, things don't actually work. For some reason, this weird set of labels now appears on the search page with the text "Alias Sesar Number Owner Collector Collectio Date" and I'm not really sure why it's there
Also, the RockType doesn't actually populate with all of the rocktypes in the list. Not really sure how I do this.