MineralAttribute: Making revision 275 the head revision instead of 293
Revision 293 was causing the chemical analysis page to not load. In the MineralAttribute code there was type casting to a SampleDTO and then an attempt to set an object that we passed in equal to the type casted object. This would cause an error because the object passed in was a ChemicalAnalysisDTO.