Bulk Upload - Clientside: Allow for request/display of Column to Name mapping
This is _ugly_ I know, but so is what is there already. This is just meant to get things functional so that when it comes time to put together the final ui for bulk upload all the functionality will exist and all that will be needed will be to move the components around.