Bulk Upload - Images: Relax zip requirements
Instead of requiring zips to be all of one form, allow some standard
deviations. Before we requirement 'main.xls' to be in the root of the
zip, now we look through the zip and find the first excel spreadsheet
and then where ever we found the xls, that is where we consider our
root to be. All image paths should be relative to that spreadsheet.