Exceptions: ValueNotInCollectionException should report the collection
Previously the error generated by a ValueNotInCollectionException only
stated that '<value> is not in the collection,' which wasn't very
helpful. Instead, remember the collection name so that the value and
it's collection can be referenced in the error message. The collection
names, and the error message itself were moved to locale files.