Search - Gives a custom message when no results are returned
ChemicalAnalysisListEx, SampleListEx, SubsampleListEx - Constructor can now take in a string which will be displayed as a custom message when no results are found

ListEx - Allows for custom "No Results" messages. Reloading the scrolltable will now also clear the "highlighted-row" css from all the rows. This is needed because the userSamplesList page will now highlight rows and we'd like them to unhighlight after any action.

UserSamplesList - Removed the newest/favorites/all tabs from the top right. Removed the links to samples in projects. 
On loading the table it will now check for a cookie that stores table columns, if it is there then the table will only display those columns unless the user changes the view manually.
When attemping to delete samples, if any samples were public then no samples will be deleted and the public samples will be highlighted in the table along with a message on the top of the page letting the user know that they attempted to delete public samples.

CustomTableView - Now takes in a cookie name (String) for a specific table. When selecting a view, it will save that view in a cookie for that specific table.