ChemicalAnalysisDetails - Changed the SampleName and SubsampleName text attributes to hyperlink attributes. Also passes in a custom no results message to our table widget
SubsampleDetails- Changed SampleName to a hyperlink attribute. Added custom no results message

SampleDetails- Custom no results message. If a user is creating a new subsample and went to that page by clicking a link on the SampleDetails page, then after saving that new subsample, they will be sent back to the SampleDetails page.