More tweaks to breadcrumbs.
Fixed the problems breadcrumbs was having with the image map
Added <leftside> tags to the pages in sitemap.xml to specify what the side bar should be for that specific page.
Left Side bar code has been moved to LeftColWidget (breadcrumbs calls the functions though) instead of having it in 10 different pages. 
The MySubsamples side bar now correctly displays which subsample you're viewing.
The left side bars and breadcrumbs won't load until you have permission to access the page (i.e. if you're asked to log in it'll wait until you logged in).
Fixed a few links to other pages that wouldn't be entered into the pages "history". This will help with navigating using the back and forward buttons on your browser.