DAO: Introduce the new DAO architecture
These objects allow for access to the objects stored in the database
and the hibernate queries used to access them. The goal here is no
server.impl class should be directly accessing the db, instead they
should be using the DAO objects, thereby allowing all db logic to
reside solely in the DAO objects, and not scattered everywhere.