Bulk Upload - CA: Return Subsamples Numbers per Subsample
Up until now the Subsample counts have been per Chemical Analysis.
In this way the number of old subsamples, was really the number of
Chemical Analysis wild subsamples that are old. Instead let's do this
the right way, and have the number of old subsamples be the number of
subsamples referenced in the spreadsheet that are old.