Got rid of the userID parameter for basicKML, it's not an effective way to have security for the google earth export (and causes users to need to login to export) so we'll have to think of a better way down the line.
You can now specify the # of columns you want for the CheckBoxesAttribute

You can now export to google earth from the "My Samples" page

Added the SearchChemistryAttribute, currently it displays all the elements/oxides which is too large a list. Will need to select a few standard/common ones and let the user add the others if he wishes.