Updated projectsetup with Kevin's change to the GWT content Created a new elements_and_oxides_mineral_types.sql which does the inserts into the respective tables AFTER the oxides/elements have been created
NOTE: There are still several errors in the recreate-database.sql execution. IT seems there are a few statements in the elements_and_oxides_mineral_types which refers to oxides that don't exist. However, the reason for these errors are because the oxides are not actually in the elements_and_oxides file. I don't know if they should be there or if we can just delete the lines referring to them.