removed automatic formatting on save - was getting annoying added in serialization and client side enumeration for postgis polygon (and supporting classes) updated searchlocationattribute to pass us back a polygon for the bounding box (which effectively means we can have them specify any number of points for a bounding box, i.e. they can make a bounding box of a star)
they way we do searches with postgis data is first we carry out everything with lucene, i.e. if they specify an alias we do that first
then once everything with lucene is complete we then apply a filter to the session.  what the filter does is it applies an extra where clause to the sql and in the case of the bounding box the where clause only returns samples whos location interesects with the bounding box