So the test I'm making, (testSampleSearchOnUsers) doesn't actually work yet but it outlines the basic idea of what I'm working on.
First, BooleanQuery privacyQuery is a search on data that is public or owned by the given user.

Current State: User owned data is filtered through. Public data does not work yet. I think this has to do with searching boolean data, or I poorly indexed the sample publicData field.

Second, this result is then part of a bigger BooleanQuery called fullQuery. This query joins the results of the above query with the actual query done by the user. In the test I'm doing, this will be a Date Range search. Unfortunately, I still can't get the date search to work. Not sure which of the many possible reasons is the cause of this.