Reviewed Anthony's additions and used them to fix the user filtering search.
Added two tests which reflect the way I see things going.

testPublicOrUserSamples is similar to testFindPublicSamples with the addition of combining the results of searching by user and by public.

testSampleFilteringByUserWithDates combines user search with date search. This is going to serve as the jumping point for next week's development where I want to incorporate the date into the older SampleSearch querys. Basically, incorporating everything we did this past week with the more generic older search in test so we can move it to development again.