The search interface is now in tabs.
rock types, region, and most of 'other' will be added to the constraint viewer widget on the left of the tabs. Still need to do minerals and chemistry.

Removing constraints from the constraint viewer does not yet actually remove them from the SearchSampleDTO. So technically clicking the x's does nothing right now, but that functionality will be added.

updated CheckBoxesAttribute so you can pass in the number columns you want.

added a function to object editor panel to get the save button, I needed it to change it's text from 'save' to 'set'.