Search should now have working range queries for chemical analysis/oxide analysis. However, I can't test it in test because I don't have permissions... so I just made sure everything loads in deployment.
Again, it's probably not as clean as it could be and should probably be more generic. I will work on that as time goes on.

In addition, it's currently only applicable for whatever format the min/max are in and ignores the type of range. I couldn't remember what we decided on that so it's just going to stay as is.