Fixed #126 and #217
When exporting My Samples to google earth or to the spreadsheet format, it now gets the values from all pages instead of just the first.

Fixed the google earth bugs with the incorrect lat/lng and the link the sample.

Would like a little more testing to be done on the test server though to be sure.