Changed the math around for chemical analysis.
If wt% is selected, then nothing is changed. If someone enters 5, it gets saved as 5 in the min/max fields, so 5% (plus the precision calculations)

If ppm is selected, then the number is multiplied by .00001. If 5 is entered, it gets saved as .00005, so .00005% or 5 parts per million

I think its better to do it this way than multiplying wt% values by 10,000 because if we one day add parts per billion we'd have to change all the values in the database. All the wt% values would need to be multiplied by 10,000,000 afterwards. While the current way we'd just take the ppb value and multiply it by .00000001.

Let me know if anyone thinks it should be switched to wt% = multiply by 10,000 and ppm = no change