The search sample object is now saved in the session whenever the user navigates away from the search page It is loaded when the user goes to the search page.
This allows us to use the back/forward buttons or normal navigation and not lose what was entered into the search interface.

Because we're actually saving the object, it goes through the same validation as it normally would. This means if some values are invalid, such as picking a date range with the start being after the end, then they won't be saved. Exceptions aren't being displayed because the user doesn't know we're saving it. 

Only thing that doesn't load back the values is the chemistry tab because that's going to changed soon anyway.