Added Sample Comments functionality. Still needs some confirmation dialogs and styling.
Added the code for searching by location in google earth. Isn't added to the actual UI because currently the plugin gets destroyed when you tab away from the Location tab. This is because it is put in a hidden div and that destroys the plugin. Some people have posted a few hacks to get around this, but they didn't seem viable for us. If it gets fixed one day, the code will be there.