overhauled SearchDB has been committed. I know commenting in it is lacking, I'm going to be spending the next few weeks cleaning up the code/better structuring redundancies.
To go over what's in there:

Search now has two facets:
sampleSearch -> returns sample Results
chemicalAnalysisSearch -> returns chemicalAnalysis Results

The front end can invoke whichever one it wants based upon the type of results the user wants back

The second facet is the integration of ChemicalAnalysis/Sample criteria in the search for one another.

So if you search for samples with criteria restricting ChemicalAnalysis, you will only get samples that meet those requirements. Vice versa with ChemicalAnalysis.

The only existing issue that I have detected so far is that I can't get chemicalAnalsis searches to work with Sample data yet. The reason for this is that Subsample_id is not a field of ChemicalAnalyses. While I changed that, it still didn't seem to appear in the index. Anthony, if you can fix that, I think all of the code is laid out so it will work as soon as it's indexed.