More updates to Make-Public Checkboxes of child objects force the parent objects to be selected. Checkboxes should cause objects to become public, but summary is currently incorrect. The dialogs are now setup to use maps provided by the backend but..
The Map provided by the backend for subsamples and chemical analyses doesn't seem to provide the necessary link between those objects and their parent sample. For subsamples this is not a problem since it keeps its parent sample as an attribute. The Maps are of type <long,List<E>> where long key seems to be arbitrary and not indicate which sample it's from.