#300/#196
Should be fixed now. I can't seem to reproduce the problem, but I believe it was sometimes caused by the image not loading before the resolution of the image was evaluated. Resizing the image should now occur after the image has been loaded.

Also fixed a bug I was getting when I accidently sorted by the first column in the image viewer. Since sorting by "image" makes no sense, it will sort by filename when that is clicked.