#445 - Sorting fixes! (for My Samples)
The reason some columns threw an error when attempting to sort by them, was because those columns contained many-to-many relationships which have no intuitive way to sort by. Sorting by these columns have been disabled. This includes Region, Metamorphic Grade, Minerals, References, Long/Lat.

Sorting fixed for # of subsamples, images, and analyses.

Sorting fixed for Rock Type. It was sorting by the rock type id which is a number rather than the name of the type. It now sorts by the name.