Users can now send project invites.
When a user receives an invite they are alerted on their My Projects page. A link will only display if the user has an invite and if they have an invite it says how many invites they have.

i.e. "You have 1 project invite!" / "You have 2 project invites!"

The page to accept invites is not yet implemented.

Added a "Make Subsamples public" link in the sample details page (also not yet implemented. The subsamples list needs to be converted to the DataList class first.)