Changed "View Samples" link into "View Members/Details", and main link now leads to the list of samples.
Cleaned up the My Projects page, getting rid of some links that we no longer plan on using.

Modified the samplesPublicOrUser filter to include samples that are in a project that you belong to. This needs to be looked at because the query that is generated is syntactically correct (returns the right results when put into psql command line) but it seems to return the same results for all projects (much to my frustration).

Added a migration that creates indices to help speed up the joins now used by the samplesPublicOrUser filter.