Fixed typo of indexes under the developer dropdown.
Fixed a crash when trying to sort by certain columns of the image viewer. Can now sort by collector, xrayimage specific columns are disabled because other types of images do not have these properties and it looks for a named query ending in /null rather than a property.

Got rid of the "prev" and "next" buttons that hovered over images while they were being displayed.

Fixed a bug in the chemical analysis element tables which displayed the same elementes over and over again in an odd fashion.