Bulk upload references updates.
Took off the 'Not Null' Attribute on filename in the GeoReferences table in schema. The migration was correct, so if you ran that your database is fine. If you happened to have rebuilt the database within the past day, just run this in psql:

ALTER TABLE georeference ALTER COLUMN filename DROP NOT NULL;