Make sure to run migration 03_29_10.sql
scaling the image on the image map is now saved
scaling the image on the image map is easier to see, the cursor changes and it is more obvious where to scale it from (corners)
scaling now updates the chemical analysis points along with it
cleaned up the rotating interface a bit, makes more sense but still needs to be styled
Added error handling for custom rotating an image
Fixed zooming bug that caused the screen to zoom off-center
Scale bar is now visible at the top and bottom of the image map. The length of bar is now matches the actual scale (if it says 2mm, it will show you how many pixels 2mm is on the screen)
Add the scale of the images to the sidebar.
Made the loading screen a bit more user friendly when rotating images.
Edited the confirmationDialogBox, it makes more sense now from a programmer's perspective
Added a new column to the image_on_grid table for the angle that an image is rotated, will need it later for rotating chemical analyses along with the image.